+++ engine/core/UUniqueID.cpp
@@ -5,15 +5,16 @@
 
 static std::random_device s_RandomDevice;
 static std::mt19937_64 s_Engine(s_RandomDevice());
-static std::uniform_int_distribution<uint64_t> s_UniformDistribution;
+static std::uniform_int_distribution<uint32_t> s_UniformDistribution;
 
 
 UUniqueID::UUniqueID() : m_UUID(s_UniformDistribution(s_Engine))
 {
-
+	
 }
 
-UUniqueID::UUniqueID(uint64_t uuid) : m_UUID(uuid)
+UUniqueID::UUniqueID(uint32_t uuid) : m_UUID(uuid)
 {
+
 
 }
